iotivity: code sync with iotivity upstream 1.3-rel branch(10/30)
These patch codes come from iotivity upstream 1.3-rel branch.
The commit list is as follows.

1. b65481fd523ad4b1bb5579fa40160a21ad85d9e8
    ("build: Generate iotivity_config.h in build_dir")
     - only resource folder

2. 5f568dfa93373427133a816734a652803f7f471a
    ("[IOT-2854][IOT-2858] anon ciphersuite disable")

3.  5fdb8a11ec9ac563f3f8e22e6ec7d1d6db4445a7
    ("[IOT-2826][IOT-2851]dat file with zero rownerid")

4. 89370a693744ba42733aeffe2880904da75b6102
    ("IOT-2826 Update *.dat files for samples")

5. 63f82bf1398a041cf52a6af96b8d83621eeb48c6
    ("Return one error response from RD server.")

6. f92bf07ea302bc55c4e19832594cc5c850c8e825
    ("IOT-2848 Fix for intermitent read error")

7. 1dfc200dac27e3bc83448f269619807791e32406
    ("Fixing Build Errors after applying -Werror")

+++ external/iotivity/iotivity_1.3-rel/resource/csdk/connectivity/src/camessagehandler.c
@@ -1507,11 +1507,11 @@ static void CALogPDUInfo(const CAData_t *data, const coap_pdu_t *pdu)
     OIC_LOG(DEBUG, ANALYZER_TAG, "CoAP Header (+ 0xFF)");
     OIC_LOG_BUFFER(DEBUG, ANALYZER_TAG,  (const uint8_t *) pdu->transport_hdr,
                    pdu->length - payloadLen);
-    OIC_LOG_V(DEBUG, ANALYZER_TAG, "CoAP Header size = [%lu]", pdu->length - payloadLen);
+    OIC_LOG_V(DEBUG, ANALYZER_TAG, "CoAP Header size = [%" PRIuPTR "]", (size_t) pdu->length - payloadLen);
 
     OIC_LOG_V(DEBUG, ANALYZER_TAG, "CoAP Payload");
     //OIC_LOG_BUFFER(DEBUG, ANALYZER_TAG, pdu->data, payloadLen);
-    OIC_LOG_V(DEBUG, ANALYZER_TAG, "CoAP Payload Size = [%lu]", payloadLen);
+    OIC_LOG_V(DEBUG, ANALYZER_TAG, "CoAP Payload Size = [%" PRIuPTR "]", payloadLen);
     OIC_LOG(DEBUG, ANALYZER_TAG, "=================================================");
     OIC_TRACE_END();
 }
